Rework jest config
- Switch from using .cjs to .ts (as ESM module) for all jest configs
- Move custom matchers to root from pointer example type guard tests
- Switch root package to ESM
