--- jest.config.js
@@ -1,4 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  projects: ["<rootDir>/packages/*"],
-};
