+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "private": true,
   "workspaces": [
     "packages/*"
@@ -13,10 +14,13 @@
   },
   "devDependencies": {
     "@hyperjump/json-schema": "1.6.7",
+    "@types/jest": "^29.5.14",
+    "chalk": "^5.4.1",
     "concurrently": "^8.2.2",
     "jest": "^29.7.0",
     "lerna": "^8.0.2",
     "nodemon": "^3.0.2",
+    "ts-jest": "^29.2.5",
     "tsx": "^4.16.2"
   },
   "dependencies": {
